pyproject.toml
[tool.poetry]
name = "smart_pokedex"
version = "0.0.0"
description = "A CLI tool that identifies Pokemon from images and provides basic information using PokeAPI."
authors = ["Maciej Krakowiak"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.32.3"
tensorflow = "^2.18.0"
scikit-learn = "^1.5.2"
pyyaml = "^6.0.2"
pydot = "^3.0.2"
graphviz = "^0.20.3"
matplotlib = "^3.9.2"
seaborn = "^0.13.2"
opencv-python = "^4.10.0.84"

[tool.poetry.dev-dependencies]
requests = "^2.32.3"
pillow = "^11.0.0"

[tool.poetry.group.dev.dependencies]
pre-commit = "^4.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
